Improve code organization and documentation
